backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymssql==2.2.11
snowflake-connector-python==3.6.0
python-dotenv==1.0.0
openai==1.6.1
pydantic==2.5.3
numpy==1.26.2
pandas==2.1.4
scipy==1.11.4
